+++ lib/knx_parser.dart
@@ -9,10 +9,12 @@ export 'src/parser/knx_project_parser.dart';
 
 // Models
 export 'src/models/knx_project.dart';
+export 'src/models/knx_flat_project.dart';
 export 'src/models/project_info.dart';
 export 'src/models/installation.dart';
 export 'src/models/topology.dart';
 export 'src/models/group_address.dart';
+export 'src/models/device_instance.dart';
 export 'src/models/location.dart';
 export 'src/models/datapoint_type.dart';
 export 'src/models/knx_keys.dart';